requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
fastf1
pymongo
dotenv
xgboost
pyaml
